[UNR-2509][MS] Moving RPCServiceTest.cpp to Private/Tests. - Removing SPATIALGDK_API from various classes/functions. - moving TestDefinitions to Public/Tests.
