+++ SpatialGDK/Source/SpatialGDK/Public/Utils/RPCRingBuffer.h
@@ -46,22 +46,22 @@ struct RPCRingBufferDescriptor
 namespace RPCRingBufferUtils
 {
 
-SPATIALGDK_API Worker_ComponentId GetRingBufferComponentId(ERPCType Type);
-SPATIALGDK_API RPCRingBufferDescriptor GetRingBufferDescriptor(ERPCType Type);
+Worker_ComponentId GetRingBufferComponentId(ERPCType Type);
+RPCRingBufferDescriptor GetRingBufferDescriptor(ERPCType Type);
 uint32 GetRingBufferSize(ERPCType Type);
 
 Worker_ComponentId GetAckComponentId(ERPCType Type);
 Schema_FieldId GetAckFieldId(ERPCType Type);
 
-SPATIALGDK_API Schema_FieldId GetInitiallyPresentMulticastRPCsCountFieldId();
+Schema_FieldId GetInitiallyPresentMulticastRPCsCountFieldId();
 
 bool ShouldQueueOverflowed(ERPCType Type);
 
 void ReadBufferFromSchema(Schema_Object* SchemaObject, RPCRingBuffer& OutBuffer);
-SPATIALGDK_API void ReadAckFromSchema(const Schema_Object* SchemaObject, ERPCType Type, uint64& OutAck);
+void ReadAckFromSchema(const Schema_Object* SchemaObject, ERPCType Type, uint64& OutAck);
 
-SPATIALGDK_API void WriteRPCToSchema(Schema_Object* SchemaObject, ERPCType Type, uint64 RPCId, const RPCPayload& Payload);
-SPATIALGDK_API void WriteAckToSchema(Schema_Object* SchemaObject, ERPCType Type, uint64 Ack);
+void WriteRPCToSchema(Schema_Object* SchemaObject, ERPCType Type, uint64 RPCId, const RPCPayload& Payload);
+void WriteAckToSchema(Schema_Object* SchemaObject, ERPCType Type, uint64 Ack);
 
 void MoveLastSentIdToInitiallyPresentCount(Schema_Object* SchemaObject, uint64 LastSentId);
