+++ SpatialGDK/Source/SpatialGDKTests/Examples/SpatialGDKExampleTest.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) Improbable Worlds Ltd, All Rights Reserved
 
-#include "TestDefinitions.h"
+#include "Tests/TestDefinitions.h"
 
 #include "HAL/PlatformFilemanager.h"
 #include "Misc/ScopeTryLock.h"
